// src/generated/api/course/models/getCourseResponseContent.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * CourseService
 * OpenAPI spec version: 1.0.0
 */
import type { CourseStats } from "./courseStats";
import type { CourseTagStats } from "./courseTagStats";

export interface GetCourseResponseContent {
	id: string;
	name: string;
	stats: CourseStats;
	tag_stats: CourseTagStats;
}